working version

+++ remove-archived/src/main/scala/helpers/ArchiveHunterFound.scala
@@ -2,5 +2,5 @@ package helpers
 
 trait ArchiveHunterLookupResult
 
-case class ArchiveHunterFound(archiveHunterId:String, archiveHunterCollection:String, size:Option[Int], beenDeleted:Boolean) extends ArchiveHunterLookupResult
+case class ArchiveHunterFound(archiveHunterId:String, archiveHunterCollection:String, size:Option[Long], beenDeleted:Boolean) extends ArchiveHunterLookupResult
 case object ArchiveHunterNotFound extends ArchiveHunterLookupResult
